Dead End: Switching between final evaluators makes no sense.
Parametric polymorphism of the object language term is lost at its
topmost evaluation.  That is, writing "term :: String" forces "term" to
be evaluated by the "UTLCTerm String" instance, and all functions
belonging to this instance return a String. Therefore, we cannot hope to
switch evaluators (here typeclass instances) during evaluation since the
return type is already fixed from the top-level type. Essentially, what
are required to implement switching from one evaluator to another during
evaluation is for the whole top-level type to be a dependent type. More
troublesome still is that the desired dependent type may not be known at
the top level as the value on which it depends may be a property of a
subterm.

More concretely, I realized I was trying to do something that looks like
this:

class Foo a where
  p :: a -> a -> Bool
  f :: a -> a -> a

newtype A a = A a
newtype B a = B a

instance Foo (A a) where
  p x y = ...
  -- The E1 below has concrete type (A a) bound on the previous line.
  f x y = E1

instance Foo (B a) where
  p x y = ...
  -- The E2 below has concrete type (B a) bound on the previous line.
  -- However, I wanted to switch to using the Foo (A a) instance for
  -- further recursive evaluation depending on the data x and y,
  -- which forces E2 to be a sum type (ie Either) or dependently typed.
  f x y = E2
  -- The type annotation would have to look something like this:
  -- f x y = E2 :: Cond (p x y) (A a) (B a)
  -- which afaik Haskell does not support. Perhaps theres an extension for
  -- dependent types, though.

It may be that there is no escaping the use of dependent types here, but
maybe all hope is not lost. One possibility to consider is defining
evaluators as multiparameter type classes, where the types of inputs and
outputs are free to differ. This could also generalize to the case of
multiple input/output sorts. It sure does seem, though, that the
functions in the instance that makes a choice about which evaluator to
continue with must be dependently typed.